refactor: refactor machine settings panel
